# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: CafeOre POS API
  version: 1.0.0
  description: 珈琲・俺POSシステムのAPI仕様
servers:
  - url: http://localhost:8080
    description: 開発環境
paths:
...
  /api/caos/lanes/{dripper}:
    put:
      summary: CaOS のドリッパーの担当者を替える（交代）
      description: |
        そのドリッパーの今日の担当者を替える。押したらその場で替える（抽出中かどうかは見ない。待機のカードもそのまま残す）。
        name が空なら担当者なし（senior は false にする）。senior は画面が sohosai-shift の上級生の名簿（seniors）で判定した値で、そのまま持つ。
        変えたら今日の担当者の全部を /api/ws/orders の {"type":"caos_lanes"} で全部の画面に配り、ほかのインスタンスへは DB の通知（caos_lanes_changed）で知らせる。
      operationId: putCaosLane
      tags:
        - caos
      parameters:
        - name: dripper
          in: path
          required: true
          description: ドリッパーの番号（1〜6）
          schema:
            type: integer
            minimum: 1
            maximum: 6
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaosLaneUpdateRequest'
      responses:
        '200':
          description: 替えた。今日の担当者の全部
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaosLanes'
        '400':
          description: 形の違うリクエスト
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/caos/lanes/swap:
    post:
      summary: CaOS の 2 つのドリッパーの担当者を入れ替える
      description: |
        2 つのドリッパーの今日の担当者（名前と上級生か）を 1 つのトランザクションで入れ替える。カードは動かさない。
        配信は PUT /api/caos/lanes/{dripper} と同じ。
      operationId: swapCaosLanes
      tags:
        - caos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaosLaneSwapRequest'
      responses:
        '200':
          description: 入れ替えた。今日の担当者の全部
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaosLanes'
        '400':
          description: 形の違うリクエスト（同じドリッパーどうし、など）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/cashier-state:
    get:
      summary: レジ状態取得
      description: レジが編集中の注文と直前に確定した注文の ID。まだ一度も同期されていなければ 404。
      operationId: getCashierState
      tags:
        - system
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashierStateResponse'
        '404':
          description: まだレジ状態が無い
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: レジ状態更新
      description: 単一のレジ状態を丸ごと置き換える。成功すると WebSocket で全クライアントへ配信される。
      operationId: updateCashierState
      tags:
        - system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashierStateUpdateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashierStateResponse'
        '400':
          description: editting_order に必須のキーが無い、または型が違う
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    CashierStateResponse:
      type: object
      required:
        - editting_order
        - submitted_order_id
        - updated_at
      properties:
        editting_order:
          type: object
          additionalProperties: true
          description: レジで編集中の注文。フロントの orderSchema の JSON をそのまま保持し、サーバーは上の階層のキーと型を確かめる以外は中身を解釈しない
        submitted_order_id:
          type: string
          format: uuid
          nullable: true
          description: 直前に確定した注文の ID。編集中は null
        updated_at:
          type: string
          format: date-time
    CashierStateUpdateRequest:
      type: object
      required:
        - editting_order
        - submitted_order_id
      properties:
        editting_order:
          type: object
          additionalProperties: true
        submitted_order_id:
          type: string
          format: uuid
          nullable: true
...
    CaosLane:
      type: object
      description: ドリッパーの今日の担当者。担当者がいなければ name が空で senior は false
      required: [dripper, name, senior, updated_at]
      properties:
        dripper:
          type: integer
          minimum: 1
          maximum: 6
        name:
          type: string
          description: 担当者の名前（前後の空白を落としたもの）。空なら担当者なし
        senior:
          type: boolean
          description: 上級生（限定を淹れられる）か。交代した時点で画面が sohosai-shift の名簿で判定した値
        updated_at:
          type: string
          format: date-time
          nullable: true
          description: 最後に替えた時刻。今日まだ替えていなければ null
    CaosLanes:
      type: object
      description: 今日（日本時間）のドリッパー 1〜6 の担当者。lanes は 6 つ全部を番号の順に持つ
      required: [day, lanes]
      properties:
        day:
          type: string
          description: 日本時間の日付（YYYY-MM-DD）。画面はこの日が今日のときだけ使う
          example: "2026-11-03"
        lanes:
          type: array
          items:
            $ref: '#/components/schemas/CaosLane'
    CaosLaneUpdateRequest:
      type: object
      required: [name, senior]
      properties:
        name:
          type: string
          maxLength: 40
          description: 担当者の名前（前後の空白は落とす）。空なら担当者なし
        senior:
          type: boolean
          description: 上級生（限定を淹れられる）か。画面が sohosai-shift の名簿（seniors）で判定して送る。name が空なら無視して false
    CaosLaneSwapRequest:
      type: object
      required: [first, second]
      properties:
        first:
          type: integer
          minimum: 1
          maximum: 6
        second:
          type: integer
          minimum: 1
          maximum: 6
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Invalid order ID format"
...